requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-autocomplete-light==3.8.2
django-registration-redux==2.8
django-versatileimagefield==2.0
googlemaps==4.4.2
idna==3.2
Pillow==8.3.1

python-magic==0.4.24
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
